Fix dimension issue in DistillationLoss and Train.

+++ src/loss/DistillationLoss.py
@@ -51,6 +51,6 @@ class DistillationLoss:
             labels_original)
         loss_teacher = self.cel(
             predictions_student_teacher,
-            labels_teacher.long())
+            labels_teacher)
 
         return self.alpha * loss_teacher + (1 - self.alpha) * loss_original
